refac(ui): split message module
split message module for clarity.
revault is a new module which will stock user role
transaction kinds and revault constants or enums.
Context is moved to the module using it: view.
Menu is moved in its own module in ui tree.
