+++ src/ui/menu.rs
@@ -1,0 +1,7 @@
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub enum Menu {
+    Home,
+    History,
+    Network,
+    Send,
+}
