+++ src/ui/view/vault.rs
@@ -4,7 +4,8 @@ use iced::{container, scrollable, Align, Column, Container, Element, Length, Row
 use crate::ui::{
     color,
     component::{badge, button, card, separation, text},
-    message::{Context, Message},
+    message::Message,
+    view::Context,
 };
 
 use crate::revaultd::model::{BroadcastedTransaction, Vault, VaultTransactions};
